Major Frontend project refactoring (resolves #95)
