+++ frontend/components/blocks/BlocksRow.js
@@ -2,7 +2,6 @@ import Link from "next/link";
 
 import { Row, Col } from "react-bootstrap";
 
-import { DataConsumer } from "../utils/DataProvider";
 import Timer from "../utils/Timer";
 
 const BlocksRow = props => (
