+++ frontend/components/transactions/TransactionsHeader.js
@@ -1,5 +1,3 @@
-import Link from "next/link";
-
 import { Row, Col } from "react-bootstrap";
 
 import { DataConsumer } from "../utils/DataProvider";
