+++ frontend/libraries/hexstring.ts
