+++ frontend/pages/transactions/index.js
@@ -1,6 +1,6 @@
 import Head from "next/head";
 
-import Content from "../../components/Content";
+import Content from "../../components/utils/Content";
 import Transactions from "../../components/transactions/Transactions";
 
 export default () => (
